Modify behaviour of BelongsToRelationship#setCanonicalRecord to first remove canonicalState before adding new canonical state
We have observed that when saving a record, if the request responds with
200 status and state that differs from what was sent, then the
(canonical) changes to belong-to relationships are not applied as
expected.

The `canonicalState` is set correctly but `inverseRecord` is never
modified. We believe this is because in the prior implementation
flushCanonical would never be called in this case.